+++ index.js
@@ -2,8 +2,8 @@
 
 const mpath = require('mpath');
 
-module.exports = function mongooseLeanGetters(schema) {
-  const fn = applyGettersMiddleware(schema);
+module.exports = function mongooseLeanGetters(schema, options) {
+  const fn = applyGettersMiddleware(schema, options);
   // Use `pre('find')` so this also works with `cursor()`
   // and `eachAsync()`, because those do not call `post('find')`
   schema.pre('find', function() {
@@ -31,8 +31,9 @@ module.exports = function mongooseLeanGetters(schema) {
   schema.post('findOneAndReplace', fn);
 };
 
-function applyGettersMiddleware(schema) {
+function applyGettersMiddleware(schema, options) {
   return function(res) {
+    this._mongooseLeanGettersOptions = options || {};
     applyGetters.call(this, schema, res);
   };
 }
@@ -41,7 +42,10 @@ function applyGetters(schema, res) {
   if (res == null) {
     return;
   }
-  if (this._mongooseOptions.lean && this._mongooseOptions.lean.getters) {
+  const { defaultLeanOptions } = this._mongooseLeanGettersOptions;
+  const shouldCallGetters = this._mongooseOptions?.lean?.getters ?? defaultLeanOptions?.getters ?? false;
+
+  if (shouldCallGetters) {
     if (Array.isArray(res)) {
       const len = res.length;
       for (let i = 0; i < len; ++i) {
@@ -70,7 +74,9 @@ function getSchemaForDoc(schema, res) {
       break;
     }
   }
-  return childSchema;
+
+  // If no discriminator schema found, return the root schema (#39)
+  return childSchema || schema;
 }
 
 function applyGettersToDoc(schema, doc) {
@@ -118,10 +124,23 @@ function applyGettersToDoc(schema, doc) {
       );
       return;
     }
-
-    if (schematype.$isSingleNested) {
-      applyGettersToDoc.call(this, schematype.schema, pathVal);
-      return;
+    
+    const pathExists = mpath.has(path, doc);
+    if (pathExists) {
+      const val = schematype.applyGetters(mpath.get(path, doc), doc, true);
+      if (val && schematype.$isMongooseArray && !schematype.$isMongooseDocumentArray) {
+        mpath.set(
+          path,
+          val.map(subdoc => {
+            return schematype.caster.applyGetters(subdoc, doc);
+          }),
+          doc
+        );
+      } if (val && schematype.$isSingleNested) {
+        applyGettersToDoc.call(this, schematype.schema, pathVal);
+      } else {
+        mpath.set(path, val, doc);
+      }
     }
 
     mpath.set(path, schematype.applyGetters(pathVal, doc, true), doc);